fixes update quota stuck resources problem
Fixes #27. With this commit we now check in a loop if the subnamespace has been properly and fully updated before doing the entire updatequota operation. In essence this makes the reconciliation of the subnamespaces involved in the updatequota to be sequential instead of random, which caused an issue with updatequota failing.

Signed-off-by: liadlauber <liad6500@gmail.com>