Added support for resourcepool migration
Fixes #15. With this PR it is now supported to migrate from and to resourcepool.
In addition, in case of failure during the migration, the migrationhierarchy object will get an Error phase with the relevant error in the object status.
Moreover, this PR fixes ResourcePool annotations bug on namespace controller, and validates Upper ResourcePool cannot be created without a quota.

Signed-off-by: liadlauber <liad6500@gmail.com>